# app/api-docs/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Awesome RAG Forge API",
    "version": "0.1.0",
    "description": "Testing/client HTTP surface for approved RAG chat, approved context/collections/document download, approved harness display, local Ollama setup helpers, and one narrow feedback write (POST /api/feedback can only create a RagFeedback row). Knowledge and harness management writes are MCP-only: this API cannot create, edit, approve, reject, archive, or delete RAG knowledge or harness rules — see docs/architecture.md. Generate a client in any language from this spec with https://openapi-generator.tech."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this deployment"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "APP_API_KEY, sent as `Authorization: Bearer <key>`. Only enforced when APP_API_KEY is configured server-side; required (fail-closed) on Vercel deployments. See lib/testing-api-auth.ts."
      }
    }
  },
  "paths": {
...
    "/api/rag/harness": {
      "get": {
        "summary": "Testing UI: approved assistant identity and harness",
        "description": "Read-only testing-surface endpoint for showing the approved assistant identity, capabilities, and restrictions. Harness proposal and management stay outside this API; pending review is handled through MCP or the guarded local /review page.",
        "tags": [
          "Harness"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Harness configuration retrieved"
          },
          "401": {
            "description": "Missing/invalid API key (only when APP_API_KEY is configured)"
          },
          "404": {
            "description": "Testing surface disabled"
          },
          "503": {
            "description": "Database unreachable"
          }
        }
      }
    },
    "/api/rag/documents/{id}/download": {
      "get": {
        "summary": "Download an approved document's stored original file",
        "description": "Requires the document to be status APPROVED, audience EXTERNAL, visibility CHAT, inside an external chat-visible collection, with a non-null storageKey. Redirects to a short-lived presigned URL by default, or returns it as JSON with ?format=json.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "json"
              ]
            },
            "description": "Pass \"json\" to receive { ok, url } instead of a 307 redirect."
          }
        ],
        "responses": {
          "200": {
            "description": "Presigned URL returned as JSON (only with ?format=json)"
          },
          "307": {
            "description": "Redirect to a short-lived presigned download URL"
          },
          "401": {
            "description": "Missing/invalid API key (only when APP_API_KEY is configured)"
          },
          "404": {
            "description": "Testing surface disabled, or no downloadable file for this document"
          },
          "503": {
            "description": "Database unreachable"
          }
        }
      }
    },
...
    "/api/rag/collections": {
      "get": {
        "summary": "List approved collections",
        "tags": [
          "Collections"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collections retrieved"
          },
          "401": {
            "description": "Missing/invalid API key (only when APP_API_KEY is configured)"
          },
          "404": {
            "description": "Testing surface disabled"
          },
          "503": {
            "description": "Database unreachable"
          }
        }
      }
    },
    "/api/rag/collections/{collectionId}": {
      "get": {
        "summary": "Paginated detail for one approved collection",
        "tags": [
          "Collections"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection detail retrieved"
          },
          "401": {
            "description": "Missing/invalid API key (only when APP_API_KEY is configured)"
          },
          "404": {
            "description": "Testing surface disabled, or collection not found"
          },
          "503": {
            "description": "Database unreachable"
          }
        }
      }
    },
...
  },
...
}